+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Corrections :
     - Optimisation et correction de l'algorithme de contrôle d'accès.
     - Erreur des noms de champ barre des membres dans le pied de page.
+    - Génération de l'image tag, amélioration du code et du message d'erreur.
 
 ## version 10.2.00
 - Mise à jour :